# backend/docs/swagger.yaml
basePath: /
definitions:
  gorm.DeletedAt:
    properties:
      time:
        type: string
      valid:
        description: Valid is true if Time is not NULL
        type: boolean
    type: object
...
  handlers.CreateCustomerAssetRequest:
    description: Datos para crear un nuevo bien del cliente
    properties:
      assetId:
        example: 1
        type: integer
      creditRequestId:
        example: 1
        type: integer
      customerId:
        example: 1
        type: integer
      description:
        example: Vehículo Toyota Corolla 2020
        type: string
      marketValue:
        example: 50000000
        type: number
    type: object
...
  handlers.UpdateCustomerAssetRequest:
    description: Datos para actualizar un bien del cliente existente
    properties:
      assetId:
        example: 2
        type: integer
      creditRequestId:
        example: 1
        type: integer
      customerId:
        example: 1
        type: integer
      description:
        example: Vehículo Toyota Corolla 2021
        type: string
      marketValue:
        example: 55000000
        type: number
    type: object
...
  models.CustomerAsset:
    properties:
      assetId:
        type: integer
      createdAt:
        type: string
      creditRequestId:
        type: integer
      customerId:
        type: integer
      deletedAt:
        $ref: '#/definitions/gorm.DeletedAt'
      description:
        type: string
      id:
        type: integer
      marketValue:
        type: number
      status:
        type: boolean
      updatedAt:
        type: string
    type: object
...
host: localhost:4000
info:
  contact:
    email: support@example.com
    name: API Support
  description: API para gestión de solicitudes de crédito y evaluación de riesgo crediticio.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  title: Credit Risk Assessment API
  version: "1.0"
paths:
...
  /customer-assets:
    get:
      consumes:
      - application/json
      description: Retorna una lista de todos los bienes de clientes, opcionalmente
        filtrados por solicitud de crédito
      parameters:
      - description: ID de la solicitud de crédito para filtrar
        in: query
        name: creditRequestId
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Lista de bienes de clientes
          schema:
            items:
              $ref: '#/definitions/models.CustomerAsset'
            type: array
        "400":
          description: creditRequestId inválido
          schema:
            type: string
        "404":
          description: Solicitud de crédito no encontrada
          schema:
            type: string
        "500":
          description: Error interno del servidor
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Obtener todos los bienes de clientes
      tags:
      - Customer Assets
    post:
      consumes:
      - application/json
      description: Crea un nuevo bien asociado a un cliente y solicitud de crédito
      parameters:
      - description: Datos del bien del cliente
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handlers.CreateCustomerAssetRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Bien creado exitosamente
          schema:
            $ref: '#/definitions/models.CustomerAsset'
        "400":
          description: Solicitud inválida
          schema:
            type: string
        "404":
          description: Cliente, activo o solicitud no encontrada
          schema:
            type: string
        "500":
          description: Error interno del servidor
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Crear un nuevo bien del cliente
      tags:
      - Customer Assets
  /customer-assets/{id}:
    delete:
      consumes:
      - application/json
      description: Elimina un bien del cliente del sistema
      parameters:
      - description: ID del bien del cliente
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "204":
          description: Bien eliminado exitosamente
        "400":
          description: ID inválido
          schema:
            type: string
        "404":
          description: Bien no encontrado
          schema:
            type: string
        "500":
          description: Error interno del servidor
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Eliminar un bien del cliente
      tags:
      - Customer Assets
    put:
      consumes:
      - application/json
      description: Actualiza los datos de un bien del cliente existente
      parameters:
      - description: ID del bien del cliente
        in: path
        name: id
        required: true
        type: integer
      - description: Datos actualizados del bien
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handlers.UpdateCustomerAssetRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Bien actualizado exitosamente
          schema:
            $ref: '#/definitions/models.CustomerAsset'
        "400":
          description: Solicitud inválida
          schema:
            type: string
        "404":
          description: Bien no encontrado
          schema:
            type: string
        "500":
          description: Error interno del servidor
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Actualizar un bien del cliente
      tags:
      - Customer Assets
...
securityDefinitions:
  BearerAuth:
    description: 'Ingresa el token JWT con el prefijo Bearer. Ejemplo: "Bearer {token}"'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"